requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langsmith>=0.1.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp>=1.27.0
opentelemetry-exporter-otlp-proto-grpc>=1.27.0
nemoguardrails>=0.9.0
guardrails-ai>=0.5.0
httpx>=0.27.0
asyncpg>=0.29.0
psycopg>=3.1.0
structlog>=24.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0
redis>=5.0.0
giskard>=2.14.0
mlflow>=2.16.0
prometheus-client>=0.20.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
python-dotenv>=1.0.0
